Refine modular app wiring
