+++ src/hooks/useLocalStorage.ts
@@ -1,0 +1,24 @@
+import { type Dispatch, type SetStateAction, useEffect, useState } from "react";
+
+export function useLocalStorage<T>(key: string, initial: T): [T, Dispatch<SetStateAction<T>>] {
+  const [state, setState] = useState<T>(() => {
+    if (typeof window === "undefined") return initial;
+    try {
+      const raw = window.localStorage.getItem(key);
+      return raw ? (JSON.parse(raw) as T) : initial;
+    } catch {
+      return initial;
+    }
+  });
+
+  useEffect(() => {
+    if (typeof window === "undefined") return;
+    try {
+      window.localStorage.setItem(key, JSON.stringify(state));
+    } catch {
+      // ignore write errors
+    }
+  }, [key, state]);
+
+  return [state, setState];
+}
